src/requirements.txt
arcade==3.0.0.dev39
pylint
pyglet
psutil
sounddevice
userpaths
Babel
duty
optimize-images
cx-Freeze
python-stopwatch